+++ vendored_parsers/tree-sitter-latex/Package.swift
@@ -3,6 +3,7 @@ import PackageDescription
 
 let package = Package(
     name: "TreeSitterLatex",
+    platforms: [.macOS(.v10_13), .iOS(.v11)],
     products: [
         .library(name: "TreeSitterLatex", targets: ["TreeSitterLatex"]),
     ],
@@ -11,26 +12,37 @@ let package = Package(
         .target(name: "TreeSitterLatex",
                 path: ".",
                 exclude: [
-                    "binding.gyp",
-                    "bindings",
-                    "benches",
-                    ".github",
-                    "examples",
                     "Cargo.toml",
-                    "test",
-                    "grammar.js",
-                    "LICENSE",
                     "Makefile",
+                    "binding.gyp",
+                    "bindings/c",
+                    "bindings/go",
+                    "bindings/node",
+                    "bindings/python",
+                    "bindings/rust",
+                    "prebuilds",
+                    "grammar.js",
                     "package.json",
-                    "README.md",
-                    "src/grammar.json",
-                    "src/node-types.json",
+                    "package-lock.json",
+                    "pyproject.toml",
+                    "setup.py",
+                    "test",
+                    "examples",
+                    ".editorconfig",
+                    ".github",
+                    ".gitignore",
+                    ".gitattributes",
+                    ".gitmodules",
                 ],
                 sources: [
                     "src/parser.c",
                     "src/scanner.c",
                 ],
+                resources: [
+                    .copy("queries")
+                ],
                 publicHeadersPath: "bindings/swift",
                 cSettings: [.headerSearchPath("src")])
-    ]
+    ],
+    cLanguageStandard: .c11
 )