+++ vendored_parsers/tree-sitter-latex/bindings/go/binding.go
@@ -1,0 +1,13 @@
+package tree_sitter_latex
+
+// #cgo CFLAGS: -std=c11 -fPIC
+// #include "../../src/parser.c"
+// #include "../../src/scanner.c"
+import "C"
+
+import "unsafe"
+
+// Get the tree-sitter Language for this grammar.
+func Language() unsafe.Pointer {
+	return unsafe.Pointer(C.tree_sitter_latex())
+}