+++ vendored_parsers/tree-sitter-latex/src/scanner.c
@@ -9,7 +9,13 @@ enum TokenType {
   TRIVIA_RAW_ENV_VERBATIM,
   TRIVIA_RAW_ENV_LISTING,
   TRIVIA_RAW_ENV_MINTED,
+  TRIVIA_RAW_ENV_ASY,
+  TRIVIA_RAW_ENV_ASYDEF,
   TRIVIA_RAW_ENV_PYCODE,
+  TRIVIA_RAW_ENV_LUACODE,
+  TRIVIA_RAW_ENV_LUACODE_STAR,
+  TRIVIA_RAW_ENV_SAGESILENT,
+  TRIVIA_RAW_ENV_SAGEBLOCK,
 };
 
 static bool find_verbatim(TSLexer *lexer, const char *keyword,
@@ -92,7 +98,7 @@ bool tree_sitter_latex_external_scanner_scan(void *payload, TSLexer *lexer,
                                              const bool *valid_symbols) {
   bool found = false;
   TSSymbol type = 0xFFFF;
-  for (int i = 0; i <= TRIVIA_RAW_ENV_PYCODE; i++) {
+  for (int i = 0; i <= TRIVIA_RAW_ENV_SAGEBLOCK; i++) {
     if (valid_symbols[i]) {
       if (found) {
         return false;
@@ -117,6 +123,18 @@ bool tree_sitter_latex_external_scanner_scan(void *payload, TSLexer *lexer,
     return find_verbatim(lexer, "\\end{minted}", false);
   case TRIVIA_RAW_ENV_PYCODE:
     return find_verbatim(lexer, "\\end{pycode}", false);
+  case TRIVIA_RAW_ENV_ASY:
+    return find_verbatim(lexer, "\\end{asy}", false);
+  case TRIVIA_RAW_ENV_ASYDEF:
+    return find_verbatim(lexer, "\\end{asydef}", false);
+  case TRIVIA_RAW_ENV_LUACODE:
+    return find_verbatim(lexer, "\\end{luacode}", false);
+  case TRIVIA_RAW_ENV_LUACODE_STAR:
+    return find_verbatim(lexer, "\\end{luacode*}", false);
+  case TRIVIA_RAW_ENV_SAGESILENT:
+    return find_verbatim(lexer, "\\end{sagesilent}", false);
+  case TRIVIA_RAW_ENV_SAGEBLOCK:
+    return find_verbatim(lexer, "\\end{sageblock}", false);
   }
 
   return false;